kpmg-ai-backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.4,<3.0.0
python-dotenv==1.0.0

# LangChain & LangGraph (versions compatibles)
langchain==0.2.16
langchain-core>=0.2.39,<0.3.0
langchain-openai==0.1.23
langgraph==0.2.40
langchain-community==0.2.16

# RAG & Vector Store
chromadb==0.4.18
langchain-chroma==0.1.2

# Document Loaders (for RAG)
pypdf==3.17.4
python-docx==1.1.0
unstructured==0.11.6

# Google Docs (optionnel - si vous voulez aussi charger depuis Google Docs)
# google-api-python-client==2.100.0
# google-auth-httplib2==0.1.1
# google-auth-oauthlib==1.1.0

# Web Search
httpx==0.25.2
aiohttp==3.9.1

# Data & Visualization
pandas==2.1.3
numpy==1.26.2
plotly==5.18.0
matplotlib==3.8.2

# Export
python-pptx==0.6.23
Pillow==10.1.0

# Utilities
python-multipart==0.0.6
